Moved cleaning job and tests to projects > ingest folder
