+++ tests/unit/test_clean_ascwds_worker_data.py
@@ -6,11 +6,11 @@ from pyspark.sql.dataframe import DataFrame
 import jobs.clean_ascwds_worker_data as job
 from utils.utils import get_spark
 from tests.test_file_data import ASCWDSWorkerData
-from projects._03_independent_cqc._02_clean.unittest_data.data import (
+from projects._01_ingest.unittest_data.data import (
     ASCWDSWorkplaceData,
 )
 from tests.test_file_schemas import ASCWDSWorkerSchemas
-from projects._03_independent_cqc._02_clean.unittest_data.schemas import (
+from projects._01_ingest.unittest_data.schemas import (
     ASCWDSWorkplaceSchemas,
 )
 from utils.column_names.cleaned_data_files.ascwds_worker_cleaned import (
